+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.12
+ * add validation support
+ * add serialization support
+ * allow definition of individual attributes
+ * add resource level `headers` option
+ * `host_config` will be deprecated in 0.2.0
+ * add `Served::Attribute::Base` class
+ 
+## 0.1.11
+ * add resource level `reasource_name` option
+ * add resource level `host` option
+
+## 0.1.10
+* make `timeout` configurable per resource
+
 ## 0.1.8
 
 * fix buggy GET request template expansion defaults